Minor changes in structure of the repository
